app/requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
Babel==2.13.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-admin==1.0.4
fastapi-mail==1.4.1
fastapi-socketio==0.0.10
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multipart==0.2.4
passlib==1.7.4
pendulum==2.1.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
wsproto==1.2.0
